refactor(client): first pass overhaul
